+++ lib/routes_page.dart
@@ -1,0 +1,51 @@
+import "package:flutter/material.dart";
+
+class RoutesPage extends StatefulWidget {
+  RoutesPage({Key key, this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  RoutesPageState createState() => RoutesPageState();
+}
+
+class RoutesPageState extends State<RoutesPage> {
+  int counter = 0;
+
+  void incrementCounter() {
+    setState(() {
+      counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        leading: Builder(
+          builder: (BuildContext context) {
+            return IconButton(
+              icon: const Icon(Icons.menu),
+              onPressed: () { Scaffold.of(context).openDrawer(); },
+              tooltip: MaterialLocalizations.of(context).openAppDrawerTooltip,
+            );
+          },
+        ),
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text("Press the button below to add a favorite transit route...", style: Theme.of(context).textTheme.display1)
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: incrementCounter,
+        tooltip: 'Increment',
+        child: Icon(Icons.add),
+      ), // This trailing comma makes auto-formatting nicer for build methods.
+    );
+  }
+}
